Update for clarity
It was unclear during my first attempt using the New Campaign interface that I needed to select one or more hash files.  It appeared that I was going to add each item in the list to my new campaign. This wording would clear the intended usage up for anyone else in a similar position.